Accept zip files as input for XHarnessApkToTest

+++ src/Microsoft.DotNet.Helix/Sdk/CreateXHarnessAndroidWorkItems.cs
@@ -88,32 +88,42 @@ namespace Microsoft.DotNet.Helix.Sdk
                 return null;
             }
 
-            if (!fileSystem.GetExtension(apkPath).Equals(".apk", StringComparison.OrdinalIgnoreCase))
-            {
-                Log.LogError($"Unsupported app package type: {fileSystem.GetFileName(apkPath)}");
-                return null;
-            }
+            string extension = fileSystem.GetExtension(apkPath).ToLowerInvariant();
+            bool isAlreadyArchived = (extension == ".zip");
 
-            // Validation of any metadata specific to Android stuff goes here
-            if (!appPackage.GetRequiredMetadata(Log, MetadataNames.AndroidPackageName, out string androidPackageName))
+            if (!isAlreadyArchived && extension != ".apk")
             {
-                Log.LogError($"{MetadataNames.AndroidPackageName} metadata must be specified; this may match, but can vary from file name");
+                Log.LogError($"Unsupported payload file `{fileSystem.GetFileName(apkPath)}`; expecting .apk or .zip");
                 return null;
             }
 
             var (testTimeout, workItemTimeout, expectedExitCode, customCommands) = ParseMetadata(appPackage);
+            appPackage.TryGetMetadata(MetadataNames.AndroidPackageName, out string androidPackageName);
 
             if (customCommands == null)
             {
                 // When no user commands are specified, we add the default `android test ...` command
                 customCommands = GetDefaultCommand(appPackage, expectedExitCode);
+
+                // Validation of any metadata specific to Android stuff goes here
+                if (!string.IsNullOrEmpty(androidPackageName))
+                {
+                    Log.LogError($"{MetadataNames.AndroidPackageName} metadata must be specified when not supplying custom commands");
+                    return null;
+                }
             }
 
             string command = GetHelixCommand(appPackage, apkPath, androidPackageName, testTimeout, expectedExitCode);
 
             Log.LogMessage($"Creating work item with properties Identity: {workItemName}, Payload: {apkPath}, Command: {command}");
 
-            string workItemZip = await CreateZipArchiveOfPackageAsync(zipArchiveManager, fileSystem, workItemName, apkPath, customCommands);
+            string workItemZip = await CreateZipArchiveOfPackageAsync(
+                zipArchiveManager,
+                fileSystem,
+                workItemName,
+                isAlreadyArchived,
+                apkPath,
+                customCommands);
 
             return CreateTaskItem(workItemName, workItemZip, command, workItemTimeout);
         }
@@ -171,19 +181,29 @@ namespace Microsoft.DotNet.Helix.Sdk
             IZipArchiveManager zipArchiveManager,
             IFileSystem fileSystem,
             string workItemName,
+            bool isAlreadyArchived,
             string fileToZip,
             string injectedCommands)
         {
-            string fileName = $"xharness-apk-payload-{workItemName.ToLowerInvariant()}.zip";
-            string outputZipPath = fileSystem.PathCombine(fileSystem.GetDirectoryName(fileToZip), fileName);
+            string outputZipPath;
 
-            if (fileSystem.FileExists(outputZipPath))
+            if (isAlreadyArchived)
             {
-                Log.LogMessage($"Zip archive '{outputZipPath}' already exists, overwriting..");
-                fileSystem.DeleteFile(outputZipPath);
+                outputZipPath = fileToZip;
             }
+            else
+            {
+                string fileName = $"xharness-apk-payload-{workItemName.ToLowerInvariant()}.zip";
+                outputZipPath = fileSystem.PathCombine(fileSystem.GetDirectoryName(fileToZip), fileName);
 
-            zipArchiveManager.ArchiveFile(fileToZip, outputZipPath);
+                if (fileSystem.FileExists(outputZipPath))
+                {
+                    Log.LogMessage($"Zip archive '{outputZipPath}' already exists, overwriting..");
+                    fileSystem.DeleteFile(outputZipPath);
+                }
+
+                zipArchiveManager.ArchiveFile(fileToZip, outputZipPath);
+            }
 
             // WorkItem payloads of APKs can be reused if sent to multiple queues at once,
             // so we'll always include both scripts (very small)
